Feature(MainReport): Bind toolbar to YAML; refactor
The main report YAML handeling has been moved to it's own function for
legibility and reusability. Then the parsed data has been handled to the
toolbar partial, where it has been merged with the flag handeling to
respond well to both the presence of flags and YAML configurations,
where flags take precedence.

In the toolbar file a refactor has been included to make the formatting
more idiomatic to Ruby styles, easing the future inclusion of aditional
formats.